r/aws_simpledb_domain: Remove terraform-plugin-sdk implementation.
