--- internal/service/simpledb/domain_fw.go
@@ -1,164 +1,0 @@
-package simpledb
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/simpledb"
-	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-provider-aws/internal/conns"
-	"github.com/hashicorp/terraform-provider-aws/internal/errs"
-	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
-)
-
-func init() {
-	registerFrameworkResourceFactory(newResourceDomain)
-}
-
-// newResourceDomain instantiates a new Resource for the aws_simpledb_domain resource.
-func newResourceDomain(context.Context) (resource.ResourceWithConfigure, error) {
-	return &resourceDomain{}, nil
-}
-
-type resourceDomain struct {
-	meta *conns.AWSClient
-}
-
-// Metadata should return the full name of the resource, such as
-// examplecloud_thing.
-func (r *resourceDomain) Metadata(_ context.Context, request resource.MetadataRequest, response *resource.MetadataResponse) {
-	response.TypeName = "aws_simpledb_domain"
-}
-
-// GetSchema returns the schema for this resource.
-func (r *resourceDomain) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
-	schema := tfsdk.Schema{
-		Attributes: map[string]tfsdk.Attribute{
-			"id": {
-				Type:     types.StringType,
-				Optional: true,
-				Computed: true,
-			},
-			"name": {
-				Type:     types.StringType,
-				Required: true,
-				// TODO ForceNew:true,
-			},
-		},
-	}
-
-	return schema, nil
-}
-
-// Configure enables provider-level data or clients to be set in the
-// provider-defined Resource type.
-func (r *resourceDomain) Configure(_ context.Context, request resource.ConfigureRequest, response *resource.ConfigureResponse) {
-	if v, ok := request.ProviderData.(*conns.AWSClient); ok {
-		r.meta = v
-	}
-}
-
-// Create is called when the provider must create a new resource.
-// Config and planned state values should be read from the CreateRequest and new state values set on the CreateResponse.
-func (r *resourceDomain) Create(ctx context.Context, request resource.CreateRequest, response *resource.CreateResponse) {
-	var data resourceDomainData
-
-	response.Diagnostics.Append(request.Plan.Get(ctx, &data)...)
-
-	if response.Diagnostics.HasError() {
-		return
-	}
-
-	name := data.Name.Value
-	input := &simpledb.CreateDomainInput{
-		DomainName: aws.String(name),
-	}
-
-	_, err := r.meta.SimpleDBConn.CreateDomainWithContext(ctx, input)
-
-	if err != nil {
-		response.Diagnostics.AddError("creating SimpleDB Domain", err.Error())
-
-		return
-	}
-
-	data.ID = types.String{Value: name}
-
-	response.Diagnostics.Append(response.State.Set(ctx, &data)...)
-}
-
-// Read is called when the provider must read resource values in order to update state.
-// Planned state values should be read from the ReadRequest and new state values set on the ReadResponse.
-func (r *resourceDomain) Read(ctx context.Context, request resource.ReadRequest, response *resource.ReadResponse) {
-	var data resourceDomainData
-
-	response.Diagnostics.Append(request.State.Get(ctx, &data)...)
-
-	if response.Diagnostics.HasError() {
-		return
-	}
-
-	_, err := FindDomainByName(ctx, r.meta.SimpleDBConn, data.ID.Value)
-
-	if tfresource.NotFound(err) {
-		response.Diagnostics.Append(errs.NewResourceNotFoundWarningDiagnostic(err))
-		response.State.RemoveResource(ctx)
-
-		return
-	}
-
-	if err != nil {
-		response.Diagnostics.AddError("reading SimpleDB Domain", err.Error())
-
-		return
-	}
-
-	response.Diagnostics.Append(response.State.Set(ctx, &data)...)
-}
-
-// Update is called to update the state of the resource.
-// Config, planned state, and prior state values should be read from the UpdateRequest and new state values set on the UpdateResponse.
-func (r *resourceDomain) Update(ctx context.Context, request resource.UpdateRequest, response *resource.UpdateResponse) {
-	var data resourceDomainData
-
-	response.Diagnostics.Append(request.Plan.Get(ctx, &data)...)
-
-	if response.Diagnostics.HasError() {
-		return
-	}
-
-	response.Diagnostics.Append(response.State.Set(ctx, &data)...)
-}
-
-// Delete is called when the provider must delete the resource.
-// Config values may be read from the DeleteRequest.
-//
-// If execution completes without error, the framework will automatically call DeleteResponse.State.RemoveResource(),
-// so it can be omitted from provider logic.
-func (r *resourceDomain) Delete(ctx context.Context, request resource.DeleteRequest, response *resource.DeleteResponse) {
-	var data resourceDomainData
-
-	response.Diagnostics.Append(request.State.Get(ctx, &data)...)
-
-	if response.Diagnostics.HasError() {
-		return
-	}
-
-	_, err := r.meta.SimpleDBConn.DeleteDomainWithContext(ctx, &simpledb.DeleteDomainInput{
-		DomainName: aws.String(data.ID.Value),
-	})
-
-	if err != nil {
-		response.Diagnostics.AddError("deleting SimpleDB Domain", err.Error())
-
-		return
-	}
-}
-
-type resourceDomainData struct {
-	ID   types.String `tfsdk:"id"`
-	Name types.String `tfsdk:"name"`
-}
